storage/requirements.txt
SQLAlchemy==1.4.31
psycopg2-binary==2.9.3
redis==4.1.4
python-dotenv==0.19.2
pytest==7.1.2
pytest-mock==3.7.0
pytest-watch==4.2.0
pydantic==1.9.0
python-dateutil==2.8.2
blinker==1.8.2
click==8.1.7
flask==3.0.3
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
watchdog==4.0.2
werkzeug==3.0.4
zipp==3.20.2
